# cmd: defer stating model info until necessary